The architecture for permissions?
Legacy often has long lines of code related to perms, often making code more unreadable (probably less efficient as well).

So why not just try to move this over to a permission manager class.

Event's are also already used, so these can technically be partly used as well for perms. I dont know if that's going to be better...