Small performance improvement, and a new type of enumeration
Reduced the use for hash-something for small lists. Sets / Maps using hashes has a O(1) notation for functions like contains(). The processing time only depends on the length of the hash (and if there are duplicate hashes). But for very small lists, this actually takes more time, and collections like ArrayLists become faster.

Second thing; Just added a enumeration for all the message keys that are used in the lang manager; Just makes the code smoother.