Skips legacy BungeeCord network permission check
This change skips the legacy Bungee network check. This means that if you are able to create a Legacy Bungee Stargate, you are able to create it on the legacy bungee network.
In addition, if for some reason, something tries to create a portal without the legacy bungee flag on the legacy bungee network, an invalid permission node will be returned to hopefully prevent the action.